requirements.txt
Django==5.1
django-allauth==65.4.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
python-dotenv==1.0.1
mysqlclient==2.2.4
channels==4.1.0
python-socketio==5.11.3
python-engineio==4.9.1 